Ref: extracting server helpers.
